+++ custom_components/ecoflow_api/button.py
@@ -13,6 +13,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from .const import (
     DEVICE_TYPE_DELTA_PRO_3,
     DEVICE_TYPE_DELTA_PRO_ULTRA,
+    DEVICE_TYPE_STREAM_MICRO_INVERTER,
     DEVICE_TYPE_STREAM_ULTRA_X,
     DOMAIN,
 )
@@ -47,10 +48,13 @@ DEVICE_BUTTON_MAP = {
     DEVICE_TYPE_DELTA_PRO_3: DELTA_PRO_3_BUTTON_DEFINITIONS,
     DEVICE_TYPE_DELTA_PRO_ULTRA: DELTA_PRO_3_BUTTON_DEFINITIONS,
     DEVICE_TYPE_STREAM_ULTRA_X: STREAM_ULTRA_X_BUTTON_DEFINITIONS,
+    DEVICE_TYPE_STREAM_MICRO_INVERTER: {},
     "delta_pro_3": DELTA_PRO_3_BUTTON_DEFINITIONS,
     "delta_pro_ultra": DELTA_PRO_3_BUTTON_DEFINITIONS,
     "Delta Pro Ultra": DELTA_PRO_3_BUTTON_DEFINITIONS,
     "stream_ultra_x": STREAM_ULTRA_X_BUTTON_DEFINITIONS,
+    "stream_micro_inverter": {},
+    "Stream Microinverter": {},
 }