+++ custom_components/ecoflow_api/const.py
@@ -55,6 +55,7 @@ DEVICE_TYPE_DELTA_PRO: Final = "delta_pro"
 DEVICE_TYPE_DELTA_2: Final = "delta_2"
 DEVICE_TYPE_DELTA_2_MAX: Final = "delta_2_max"
 DEVICE_TYPE_STREAM_ULTRA_X: Final = "stream_ultra_x"
+DEVICE_TYPE_STREAM_MICRO_INVERTER: Final = "stream_micro_inverter"
 DEVICE_TYPE_SMART_PLUG: Final = "smart_plug"
 DEVICE_TYPE_POWERSTREAM_MICRO_INVERTER: Final = "powerstream_micro_inverter"
 
@@ -65,6 +66,7 @@ DEVICE_TYPES: Final = {
     DEVICE_TYPE_DELTA_2: "Delta 2",
     DEVICE_TYPE_DELTA_2_MAX: "Delta 2 Max",
     DEVICE_TYPE_STREAM_ULTRA_X: "Stream Ultra X",
+    DEVICE_TYPE_STREAM_MICRO_INVERTER: "Stream Microinverter",
     DEVICE_TYPE_SMART_PLUG: "Smart Plug S401",
     DEVICE_TYPE_POWERSTREAM_MICRO_INVERTER: "Powerstream Micro Inverter",
 }